refactor(devtools): import the jest fixtures from the package directly
Tests took the fixtures through a per-package jest/render re-export, so reading a
test meant two hops to reach the real helper. They now import
@support/jest-devtools-fixtures/web or /native directly, which removes five shim
files, the jest/render.native tsconfig aliases and the matching moduleNameMapper
entry. feature-flags no longer needs a jest/ folder at all.

The subpath is explicit rather than derived from the filename because jest's
resolver does not apply the react-native export condition — the same reason the
native preset maps lumen by hand.

devtools/shell keeps jest/screens.native.tsx for renderWithNavigation and
renderScreen, which need its navigation and shell context; it re-exports the
package so shell's tests still need only one import.
