--- devtools/feature-flags/jest/render/index.native.tsx
@@ -1,1 +1,0 @@
-export * from "@support/jest-devtools-fixtures/native";
