--- devtools/feature-flags/jest/render/index.tsx
@@ -1,1 +1,0 @@
-export * from "@support/jest-devtools-fixtures/web";
