+++ devtools/feature-flags/src/components/FlagEditorBottomSheet/FlagEditorBottomSheetContent.native.test.tsx
@@ -1,4 +1,9 @@
-import { render, screen, userEvent, BottomSheetWrapper } from "jest/render.native";
+import {
+  render,
+  screen,
+  userEvent,
+  BottomSheetWrapper,
+} from "@support/jest-devtools-fixtures/native";
 import { useJsonEditor, type JsonEditorPropsState } from "../../hooks/useJsonEditor";
 import type { FlagDisplayState } from "../../types";
 import { FlagEditorBottomSheetContent } from "./FlagEditorBottomSheetContent";
