+++ devtools/feature-flags/src/components/SortButton/useSortButtonViewModel.native.test.ts
@@ -1,4 +1,4 @@
-import { renderHook } from "jest/render.native";
+import { renderHook } from "@support/jest-devtools-fixtures/native";
 import type { SortCategory, SortDirection } from "../../hooks";
 import { useSortButtonViewModel } from "./useSortButtonViewModel";
