+++ devtools/feature-flags/src/components/filterFlagControl/FilterFlagControl.native.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "jest/render.native";
+import { render, screen } from "@support/jest-devtools-fixtures/native";
 import { FilterFlagControl } from "./FilterFlagControl";
 
 describe("FilterFlagControl (native)", () => {
