+++ devtools/feature-flags/src/components/flagDiffView/FlagDiffView.native.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "jest/render.native";
+import { render, screen } from "@support/jest-devtools-fixtures/native";
 import type { DiffLine } from "../../utils";
 import { FlagDiffView } from "./FlagDiffView";
