+++ devtools/feature-flags/src/components/flagEnableIndicator/FlagEnableIndicator.native.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "jest/render.native";
+import { render, screen } from "@support/jest-devtools-fixtures/native";
 import { FlagEnableIndicator } from "./FlagEnableIndicator";
 
 describe("FlagEnableIndicator (native)", () => {
