+++ devtools/feature-flags/src/components/flagEnableIndicator/FlagEnableIndicator.web.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "jest/render";
+import { render, screen } from "@support/jest-devtools-fixtures/web";
 import { FlagEnableIndicator } from "./FlagEnableIndicator";
 
 describe("FlagEnableIndicator", () => {
