+++ devtools/feature-flags/src/components/flagList/FlagList.web.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "jest/render";
+import { render, screen } from "@support/jest-devtools-fixtures/web";
 import userEvent from "@testing-library/user-event";
 import { FEATURE_FLAGS_INITIAL_STATE } from "@shared/feature-flags";
 import type { PartialFeatures } from "@shared/feature-flags";
