+++ devtools/feature-flags/src/components/flagMenu/useFlagMenuViewModel.native.test.ts
@@ -1,4 +1,4 @@
-import { renderHook, waitFor } from "jest/render.native";
+import { renderHook, waitFor } from "@support/jest-devtools-fixtures/native";
 import { FEATURE_FLAGS_INITIAL_STATE, type PartialFeatures } from "@shared/feature-flags";
 import {
   useFeatureFlagsToolActions,
