+++ devtools/feature-flags/src/components/flagRow/FlagRow.native.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen, userEvent } from "jest/render.native";
+import { render, screen, userEvent } from "@support/jest-devtools-fixtures/native";
 import { FEATURE_FLAGS_INITIAL_STATE } from "@shared/feature-flags";
 import type { FlagDisplayState } from "../../types";
 import {
