+++ devtools/feature-flags/src/components/pill/Pill.native.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "jest/render.native";
+import { render, screen } from "@support/jest-devtools-fixtures/native";
 import { Text } from "react-native";
 import { Pill } from "./Pill";
