+++ devtools/feature-flags/src/components/pill/Pill.web.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "jest/render";
+import { render, screen } from "@support/jest-devtools-fixtures/web";
 import { Pill } from "./Pill";
 
 describe("Pill", () => {
