+++ devtools/feature-flags/src/components/toolBar/ToolBar.web.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from "jest/render";
+import { render, screen } from "@support/jest-devtools-fixtures/web";
 import userEvent from "@testing-library/user-event";
 import { ToolBar } from "./ToolBar";
 import type { ToolBarInput, ToolBarFilters, ToolBarSort, ToolBarActions } from "./types";
