+++ devtools/feature-flags/src/feature-flags/FeatureFlags.web.test.tsx
@@ -1,4 +1,4 @@
-import { render } from "jest/render";
+import { render } from "@support/jest-devtools-fixtures/web";
 import { FEATURE_FLAGS_INITIAL_STATE } from "@shared/feature-flags";
 import { FeatureFlags } from "./FeatureFlags";
 import type { FeatureFlagsToolProps } from "../types";
