+++ devtools/feature-flags/tsconfig.json
@@ -1,10 +1,6 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
-        "paths": {
-            "jest/*": ["./jest/*"],
-            "jest/render.native": ["./jest/render/index.native.tsx"]
-        },
         "lib": ["ES2022", "DOM"],
         "target": "ES2022",
         "module": "ESNext",
