+++ devtools/shell/jest/screens.native.tsx
@@ -1,7 +1,7 @@
 import React, { type ComponentType, type ReactNode } from "react";
 import { render, type RenderResult } from "@support/jest-devtools-fixtures/native";
 import { NavigationContainer } from "@react-navigation/native";
-import { DevToolsShellProvider, type CategoryGroup } from "../../src/context";
+import { DevToolsShellProvider, type CategoryGroup } from "../src/context";
 
 function NavigationWrapper({ children }: { readonly children: ReactNode }) {
   return <NavigationContainer>{children}</NavigationContainer>;
