+++ devtools/transport-panel/src/components/HistoryLine/HistoryLine.native.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen, fireEvent } from "jest/render.native";
+import { render, screen, fireEvent } from "@support/jest-devtools-fixtures/native";
 import type { Envelope, MessageMap } from "@devtools/transport";
 import { HistoryLine } from "./HistoryLine.native";
