+++ support/jest-devtools-fixtures/README.md
@@ -42,18 +42,20 @@ module.exports = require("@support/jest-devtools-fixtures").createNativeJestConf
 
 ## Render fixtures
 
-Re-export the platform entry point from the package's own `jest/render` folder, so tests keep
-importing `jest/render` (web) and `jest/render.native` (native):
+Tests import the entry point for their platform directly. There is no per-package indirection:
 
 ```ts
-// jest/render/index.tsx
-export * from "@support/jest-devtools-fixtures/web";
+// FlagRow.web.test.tsx
+import { render, screen } from "@support/jest-devtools-fixtures/web";
 ```
 
 ```ts
-// jest/render/index.native.tsx
-export * from "@support/jest-devtools-fixtures/native";
+// FlagRow.native.test.tsx
+import { render, screen } from "@support/jest-devtools-fixtures/native";
 ```
+
+The subpath is explicit rather than resolved from the filename, because jest's resolver does not
+apply the `react-native` export condition — the same reason the native preset maps lumen by hand.
 
 Both entry points re-export their `@testing-library` package, so `screen`, `userEvent`, `renderHook`
 and friends come from the same import. `render` accepts a `wrapper` option for an extra provider
@@ -63,8 +65,9 @@ nested *inside* the `ThemeProvider`:
 render(<FlagEditorBottomSheetContent {...props} />, { wrapper: BottomSheetWrapper });
 ```
 
-Anything that needs a provider specific to one package — `@devtools/shell`'s `NavigationContainer`
-and `renderScreen`, for instance — stays in that package's `jest/render` file and builds on `render`.
+A provider specific to one package stays in that package and builds on `render` — see
+`devtools/shell/jest/screens.native.tsx`, which adds `NavigationContainer` and `renderScreen` and
+re-exports this package so its tests still need only one import.
 
 ## What stays in the consumer
 
@@ -72,4 +75,4 @@ and `renderScreen`, for instance — stays in that package's `jest/render` file 
   the consumer's root. `babel-jest` needs the React Native preset to parse the Flow-typed `.js`
   files shipped inside `react-native`; without it the native project fails to start.
 - Package-specific fixtures and mocks (`jest/test-utils.ts`, `jest/mocks/transport.ts`, …).
-- The `jest/*` tsconfig path mapping and jest `modulePaths`, which resolve the two above.
+- The `jest/*` tsconfig path mapping and jest `modulePaths`, which resolve those.
